+++ src/chip_control.h
@@ -4,7 +4,6 @@
 #include <stdbool.h>
 #include <stdint.h>
 #include <stddef.h>
-#include "ltc2943_sim.h"
 #include "ltc2943_config.h"
 
 /**
