+++ test/CMakeLists.txt
@@ -2,20 +2,19 @@ cmake_minimum_required(VERSION 3.14)
 
 set(TESTFILES
     main.c
+    unittests.c
 )
 
 set(TEST_MAIN unit_tests)
 set(TEST_RUNNER_PARAMS "")
 
+
 # --------------------------------------------------------------------------------
 #                         Make Tests
 # --------------------------------------------------------------------------------
 include_directories(${PROJECT_SOURCE_DIR}/../src)
-add_library(cpputest STATIC IMPORTED)
-set_property(TARGET cpputest PROPERTY
-             IMPORTED_LOCATION ${PROJECT_SOURCE_DIR}/extern/cpputest/build/lib/libCppUTest.a)
 add_executable(${TEST_MAIN} ${TESTFILES})
-target_link_libraries(${TEST_MAIN} PRIVATE ${LIBRARY_NAME} cpputest)
+target_link_libraries(${TEST_MAIN} PRIVATE ${LIBRARY_NAME} unity m)
 set_target_properties(${TEST_MAIN} PROPERTIES RUNTIME_OUTPUT_DIRECTORY ${PROJECT_BINARY_DIR})
 target_set_warnings(${TEST_MAIN} ENABLE ALL AS_ERROR ALL DISABLE Annoying) # Set warnings (if needed).
